+++ package.json
@@ -14,22 +14,22 @@
   "devDependencies": {
     "@tailwindcss/forms": "^0.3.4",
     "@tailwindcss/typography": "^0.4.1",
-    "@vitejs/plugin-vue": "^1.9.2",
-    "@vue/compiler-sfc": "^3.2.19",
-    "autoprefixer": "^10.3.6",
-    "chart.js": "^3.5.1",
+    "@vitejs/plugin-vue": "^1.10.0",
+    "@vue/compiler-sfc": "^3.2.22",
+    "autoprefixer": "^10.4.0",
+    "chart.js": "^3.6.0",
     "crypto-js": "^4.1.1",
-    "eslint": "^7.32.0",
-    "eslint-plugin-vue": "^7.18.0",
+    "eslint": "^8.3.0",
+    "eslint-plugin-vue": "^8.1.1",
     "feather-icons": "^4.28.0",
-    "highlight.js": "^11.2.0",
+    "highlight.js": "^11.3.1",
     "markdown-it": "^12.2.0",
     "mockjs": "^1.1.0",
-    "postcss": "^8.3.8",
-    "qiniu-js": "^3.3.3",
-    "tailwindcss": "^2.2.16",
-    "typescript": "^4.4.3",
-    "vite": "^2.5.10",
+    "postcss": "^8.3.11",
+    "qiniu-js": "^3.4.0",
+    "tailwindcss": "^2.2.19",
+    "typescript": "^4.5.2",
+    "vite": "^2.6.14",
     "vite-plugin-mock": "^2.9.6"
   },
   "keywords": [
