+++ src/App.vue
@@ -1,3 +1,3 @@
 <template>
-  <router-view />
+  <RouterView />
 </template>
