+++ src/components/Action.vue
