+++ src/components/tree/Tree.vue
@@ -11,7 +11,7 @@
 </template>
 
 <script lang="ts" setup>
-import Model from "/@/components/global/Model.vue";
+import Model from "/@/components/Model.vue";
 import TreeCore from "/@/components/tree/TreeCore.vue";
 
 defineProps({
