+++ src/mock/assets/resource.ts
