+++ src/mock/assets/statistics.ts
