+++ src/mock/index.ts
@@ -1,0 +1,30 @@
+import { createProdMockServer } from 'vite-plugin-mock/es/createProdMockServer';
+
+import accountMock from './system/account';
+import roleMock from './system/role';
+import userMock from './system/user';
+import groupMock from './system/group';
+import authorityMock from './system/authority';
+import regionMock from './system/region';
+
+import postsMock from './assets/posts';
+import categoryMock from './assets/category';
+import resourceMock from './assets/resource';
+import statisticsMock from './assets/statistics';
+import commentMock from './assets/comment';
+
+export function setupProdMockServer() {
+  createProdMockServer([
+    ...accountMock,
+    ...roleMock, 
+    ...userMock, 
+    ...groupMock, 
+    ...authorityMock,
+    ...regionMock,
+    ...postsMock, 
+    ...categoryMock, 
+    ...resourceMock,
+    ...statisticsMock,
+    ...commentMock
+  ]);
+}
