+++ src/views/system/Index.vue
@@ -1,5 +1,5 @@
 <template>
-  <router-view></router-view>
+  <RouterView></RouterView>
 </template>
 
 <script lang="ts" setup>
