+++ src/views/system/User.vue
@@ -406,10 +406,10 @@
 <script lang="ts" setup>
 import { onMounted, ref } from "vue";
 
-import Action from "/@/components/global/Action.vue";
-import Pagation from "/@/components/global/Pagation.vue";
-import Confirm from "/@/components/global/Confirm.vue";
-import Model from "/@/components/global/Model.vue";
+import Action from "/@/components/Action.vue";
+import Pagation from "/@/components/Pagation.vue";
+import Confirm from "/@/components/Confirm.vue";
+import Model from "/@/components/Model.vue";
 import Tree from "/@/components/tree/Tree.vue";
 
 import instance from "../../api";
